moved our demo to a demo folder
